@@ -1,6 +1,6 @@
 {
   "name": "react-native-gifted-chat",
-  "version": "0.1.4",
+  "version": "0.2.5",
   "description": "The most complete chat UI for React Native",
   "main": "index.js",
   "repository": {
@@ -27,23 +27,24 @@
     "lint": "eslint . --ext .js,.jsx"
   },
   "devDependencies": {
-    "eslint": "~3.12.2",
-    "eslint-config-airbnb": "~13.0.0",
-    "eslint-config-cooperka": "~0.2.3",
-    "eslint-plugin-import": "~2.2.0",
-    "eslint-plugin-jsx-a11y": "~2.2.3",
-    "eslint-plugin-react": "~6.8.0",
-    "eslint-plugin-react-native": "~2.2.1"
+    "eslint": "3.19.0",
+    "eslint-config-airbnb": "15.0.2",
+    "eslint-config-cooperka": "0.2.3",
+    "eslint-plugin-import": "2.6.1",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.1.0",
+    "eslint-plugin-react-native": "2.3.2"
   },
   "dependencies": {
-    "@expo/react-native-action-sheet": "github:kesha-antonov/react-native-action-sheet#1dc9afb93fad90858ccca02dcfc23c2f12afe71a",
+    "@expo/react-native-action-sheet": "cribspot/react-native-action-sheet#b422c54f26d30b9e85eb14a882ea9cec38600730",
     "md5": "2.2.1",
-    "moment": "2.17.1",
+    "moment": "2.18.1",
+    "prop-types": "15.5.10",
     "react-native-communications": "2.2.1",
     "react-native-invertible-scroll-view": "1.0.0",
     "react-native-lightbox": "oblador/react-native-lightbox#c84a8543d4511fe6a44c3d7820747c9c1bddd875",
-    "react-native-parsed-text": "0.0.16",
-    "shallowequal": "0.2.2",
-    "uuid": "3.0.1"
+    "react-native-parsed-text": "0.0.18",
+    "shallowequal": "1.0.2",
+    "uuid": "3.1.0"
   }
 }
