@@ -1,6 +1,8 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {
   View,
+  ViewPropTypes,
   StyleSheet,
 } from 'react-native';
 
@@ -41,11 +43,15 @@ export default class Message extends React.Component {
   }
 
   renderAvatar() {
-    if (this.props.user._id !== this.props.currentMessage.user._id) {
-      const avatarProps = this.getInnerComponentProps();
-      return <Avatar {...avatarProps}/>;
+    if (this.props.user._id === this.props.currentMessage.user._id && !this.props.showUserAvatar) {
+      return null;
     }
-    return null;
+    const avatarProps = this.getInnerComponentProps();
+    const { currentMessage } = avatarProps;
+    if (currentMessage.user.avatar === null) {
+      return null;
+    }
+    return <Avatar {...avatarProps} />;
   }
 
   render() {
@@ -86,7 +92,7 @@ const styles = {
 };
 
 Message.defaultProps = {
-  renderAvatar: null,
+  renderAvatar: undefined,
   renderBubble: null,
   renderDay: null,
   position: 'left',
@@ -98,16 +104,17 @@ Message.defaultProps = {
 };
 
 Message.propTypes = {
-  renderAvatar: React.PropTypes.func,
-  renderBubble: React.PropTypes.func,
-  renderDay: React.PropTypes.func,
-  position: React.PropTypes.oneOf(['left', 'right']),
-  currentMessage: React.PropTypes.object,
-  nextMessage: React.PropTypes.object,
-  previousMessage: React.PropTypes.object,
-  user: React.PropTypes.object,
-  containerStyle: React.PropTypes.shape({
-    left: View.propTypes.style,
-    right: View.propTypes.style,
+  renderAvatar: PropTypes.func,
+  showUserAvatar: PropTypes.bool,
+  renderBubble: PropTypes.func,
+  renderDay: PropTypes.func,
+  position: PropTypes.oneOf(['left', 'right']),
+  currentMessage: PropTypes.object,
+  nextMessage: PropTypes.object,
+  previousMessage: PropTypes.object,
+  user: PropTypes.object,
+  containerStyle: PropTypes.shape({
+    left: ViewPropTypes.style,
+    right: ViewPropTypes.style,
   }),
 };
