@@ -1,21 +1,20 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {
   Image,
   StyleSheet,
   View,
-  Dimensions,
+  ViewPropTypes,
 } from 'react-native';
 import Lightbox from 'react-native-lightbox';
 
 export default class MessageImage extends React.Component {
   render() {
-    const { width, height } = Dimensions.get('window');
-
     return (
       <View style={[styles.container, this.props.containerStyle]}>
         <Lightbox
           activeProps={{
-            style: [styles.imageActive, { width, height }],
+            style: styles.imageActive,
           }}
           {...this.props.lightboxProps}
         >
@@ -41,6 +40,7 @@ const styles = StyleSheet.create({
     resizeMode: 'cover',
   },
   imageActive: {
+    flex: 1,
     resizeMode: 'contain',
   },
 });
@@ -51,14 +51,12 @@ MessageImage.defaultProps = {
   },
   containerStyle: {},
   imageStyle: {},
-  imageProps: {},
-  lightboxProps: {},
 };
 
 MessageImage.propTypes = {
-  currentMessage: React.PropTypes.object,
-  containerStyle: View.propTypes.style,
+  currentMessage: PropTypes.object,
+  containerStyle: ViewPropTypes.style,
   imageStyle: Image.propTypes.style,
-  imageProps: React.PropTypes.object,
-  lightboxProps: React.PropTypes.object,
+  imageProps: PropTypes.object,
+  lightboxProps: PropTypes.object,
 };
